test: 유저예약 - 어드민예약 컨트롤러 테스트 분리
